[#16] refactor:폴더 구조 변경
History의 보관함 파일과 히스토리 파일을 구분하여 History와 Scrap으로 구조 변경
