+++ src/App.js
@@ -13,8 +13,8 @@ import Register from 'components/Auth/Register/Register';
 import HistoryPage from 'pages/HistoryPage';
 import { setTokenOnHeader } from 'api/logined';
 import { getCookie } from 'components/Shared/Cookies';
-import ScrapContent from 'components/History/ScrapContent';
-import HistoryContent from 'components/History/HistoryContent';
+import ScrapContent from 'components/History/Scrap/ScrapContent';
+import HistoryContent from 'components/History/History/HistoryContent';
 import MyPage from 'pages/MyPage';
 import ChatPage from 'pages/ChatPage';
 import ChatAuth from 'components/Chat/ChatAuth';
